Completion of basic python rewrite of the project.
Reworked the entirety of the project in python, improving the parsing of
scripture verse names in the process. Also allows for confiuguring how a
verse should be printed, whether to standard out, or into a paging
software. The patch even allows one to choose the maximum amount of
characters in a row, making it more versatile for other use cases.